Copy travel1
Copy of travel.csv file. This file has the original latitude and longitude values that I was going to use on the travel map project, but the points didn't show up correctly on the map. This was no doubt due to p5.js orientation parameters that positioned the map center off to the right and down (plus reversed y axis).